commands rework
